backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
supabase
pydantic
python-jose
passlib
bcrypt
aiosmtplib
pydantic[email]